Fix ]ghost for DM channels.

+++ commands/ghost.js
@@ -15,6 +15,10 @@ class GhostCommand extends Command {
 
     runInternal(msg, args) {
         let ghost_msg = args.slice(1).join(" ").replace("@", "");
+        if (msg.channel.type == 'dm' || msg.channel.type == 'group') {
+            msg.channel.send("My ghost powers don't work here!");
+            return;
+        }
         msg.delete();
         msg.channel.send(ghost_msg);
     }
